move trace clearance helpers to utils
