+++ lib/utils/offsetCollinearConnections.ts
@@ -7,8 +7,8 @@ import type {
   PinId,
 } from "../types/InputProblem"
 import type { Placement } from "../types/OutputLayout"
-import { getRotatedSize, rotatePinOffset } from "../utils/rotatePinOffset"
-import { getGroundedLoadPairs } from "./GroundedLoadPairSolver/getGroundedLoadPairs"
+import { getRotatedSize, rotatePinOffset } from "./rotatePinOffset"
+import { getGroundedLoadPairs } from "../solvers/GroundedLoadPairSolver/getGroundedLoadPairs"
 
 const TRACE_CLEARANCE = 0.2
 const ALIGNMENT_TOLERANCE = 1e-6
